requirements.txt
duckdb
pandas
langchain
langchain-google-genai
python-dotenv
pytest
tabulate
openpyxl
streamlit
langchain-openai
